+++ pkg/label/app.go
@@ -1,0 +1,7 @@
+package label
+
+const (
+	// AppOperatorWatching is the label added to configmaps by app-operator when
+	// it is watching for values changes.
+	AppOperatorWatching = "app-operator.giantswarm.io/watching"
+)