+++ service/controller/cluster.go
@@ -272,6 +272,9 @@ func newClusterResources(config ClusterConfig) ([]resource.Interface, error) {
 			K8sClient: config.K8sClient.K8sClient(),
 			Logger:    config.Logger,
 
+			AllowedLabels: []string{
+				label.AppOperatorWatching,
+			},
 			Name:        clusterconfigmap.Name,
 			StateGetter: clusterConfigMapGetter,
 		}